ci: fix `upload-image` action  (#2935)
* chore: move fake_image to linux_amd64


* chore: fix upload image


* fix: remove deafult project env


* chore: add FakeImage to fake_image


* chore: update e2e image name


* chore: fix e2e test


* chore: fix e2e test
